third_party/blink/web_tests/external/wpt/editing/run/forecolor_2001-last-expected.txt

This is a testharness.js-based test.
Found 8 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["stylewithcss","true"],["forecolor","#0000FF"]] "fo[o<font color=brown>bar</font>b]az" queryCommandIndeterm("forecolor") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["forecolor","#0000FF"]] "fo[o<font color=brown>bar</font>b]az" queryCommandValue("forecolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 0, 255)" but got "rgb(0, 0, 0)"
[FAIL] [["stylewithcss","false"],["forecolor","#0000FF"]] "fo[o<font color=brown>bar</font>b]az" queryCommandIndeterm("forecolor") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["forecolor","#0000FF"]] "<span style=color:brown>fo[o</span><span style=color:#0000ff>b]ar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"color:rgb(165, 42, 42)\\">fo</span><span style=\\"color:rgb(0, 0, 255)\\">obar</span>" but got "<span style=\\"color:rgb(165, 42, 42)\\">fo</span><span style=\\"color:rgb(0, 0, 255)\\">ob</span><span style=\\"color:rgb(0, 0, 255)\\">ar</span>"
[FAIL] [["stylewithcss","true"],["forecolor","#0000FF"]] "<span style=color:brown>fo[o</span><span style=color:#0000ff>b]ar</span>" queryCommandIndeterm("forecolor") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["forecolor","#0000FF"]] "<span style=color:brown>fo[o</span><span style=color:#0000ff>b]ar</span>" queryCommandIndeterm("forecolor") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["styleWithCSS","true"],["foreColor","rgba(0, 0, 255, 0.5)"],["styleWithCSS","false"],["insertText","b"]] "a[]c" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "a<span style=\\"color:rgba(0, 0, 255, 0.5)\\">b</span>c" but got "a<font color=\\"rgba(0, 0, 255, 0.5)\\">b</font>c"
[FAIL] [["styleWithCSS","true"],["foreColor","transparent"],["styleWithCSS","false"],["insertText","b"]] "a[]c" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "a<span style=\\"color:rgba(0, 0, 0, 0)\\">b</span>c" but got "a<font color=\\"rgba(0, 0, 0, 0)\\">b</font>c"
Harness: the test ran to completion.
